+++ core/src/main/scala/com/rackspace/com/papi/components/checker/step/base/StepContext.scala
@@ -19,8 +19,8 @@ import com.rackspace.com.papi.components.checker.handler.ResultHandler
 import com.rackspace.com.papi.components.checker.util.HeaderMap
 
 import net.sf.saxon.s9api.XdmValue
+import net.sf.saxon.s9api.XdmMap
 import net.sf.saxon.Configuration
-import net.sf.saxon.expr.JPConverter
 
 import collection.JavaConverters._
 
@@ -28,12 +28,12 @@ import collection.JavaConverters._
 //  Used to keep context about the current request
 //
 case class StepContext(uriLevel : Int = 0, requestHeaders : HeaderMap = new HeaderMap, handler: Option[ResultHandler] = None) {
-  def asXdmValue(c : Configuration) : XdmValue = {
+  def asXdmValue : XdmValue = {
     val headers : java.util.Map[String, java.util.List[String]] = new java.util.HashMap[String,java.util.List[String]]()
     requestHeaders.foreach { case (k : String, l : List[String]) =>  headers.put (k, l.asJava) }
     val jcontext : java.util.Map[String, Object] = new java.util.HashMap[String, Object]()
     jcontext.put("uriLevel", uriLevel.asInstanceOf[Object])
     jcontext.put("headers", headers)
-    XdmValue.wrap(JPConverter.FromMap.INSTANCE.convert(jcontext, c.getConversionContext))
+    XdmMap.makeMap(jcontext)
   }
 }
