+++ core/src/test/scala/com/rackspace/com/papi/components/checker/ValidatorWADLPlainParamSuite.scala
@@ -366,48 +366,6 @@ class ValidatorWADLPlainParamSuite extends BaseValidatorSuite {
                  false, false, true))
 
   //
-  // Like validator_XSDElementContentPlainOptMsgCode but using Saxon
-  //
-  val validator_XSDElementContentPlainOptMsgCodeS = Validator((localWADLURI,
-      <application xmlns="http://wadl.dev.java.net/2009/02"
-                   xmlns:tst="http://www.rackspace.com/repose/wadl/checker/step/test"
-                   xmlns:rax="http://docs.rackspace.com/api">
-        <grammars>
-           <include href="src/test/resources/xsd/test-urlxsd.xsd"/>
-        </grammars>
-        <resources base="https://test.api.openstack.com">
-           <resource path="/a/b">
-               <method name="PUT">
-                  <request>
-                      <representation mediaType="application/xml" element="tst:a">
-                          <param style="plain" path="tst:a/@stepType" required="true" rax:message="No stepType attribute on a" rax:code="500"/>
-                      </representation>
-                      <representation mediaType="application/json"/>
-                  </request>
-               </method>
-               <method name="POST">
-                  <request>
-                      <representation mediaType="application/xml" element="tst:e">
-                          <param style="plain" path="tst:e/tst:stepType" required="true" rax:message="no stepType on e" rax:code="501"/>
-                      </representation>
-                  </request>
-               </method>
-           </resource>
-           <resource path="/c">
-               <method name="POST">
-                  <request>
-                      <representation mediaType="application/json"/>
-                  </request>
-               </method>
-               <method name="GET"/>
-           </resource>
-        </resources>
-    </application>)
-    , TestConfig(false, false, true, true, true, 10,
-                 true, false , false, "SaxonHE", true,
-                 false, false, true))
-
-  //
   // Like validator_XSDElementContentPlainOpt but with custom rax:code
   //
   val validator_XSDElementContentPlainOptCode = Validator((localWADLURI,
@@ -1088,10 +1046,6 @@ class ValidatorWADLPlainParamSuite extends BaseValidatorSuite {
     validator_XSDElementContentPlainOptMsgCodeX.validate(request("PUT","/a/b","application/xml", goodXML_XSD2),response,chain)
   }
 
-  test ("PUT on /a/b with application/xml should succeed on validator_XSDElementContentPlainOptMsgCodeS with valid XML1") {
-    validator_XSDElementContentPlainOptMsgCodeS.validate(request("PUT","/a/b","application/xml", goodXML_XSD2),response,chain)
-  }
-
   test ("PUT on /a/b with application/xml should succeed on validator_XSDElementContentPlainOpt2 with valid XML1") {
     validator_XSDElementContentPlainOpt2.validate(request("PUT","/a/b","application/xml", goodXML_XSD2),response,chain)
   }
@@ -1122,10 +1076,6 @@ class ValidatorWADLPlainParamSuite extends BaseValidatorSuite {
 
   test ("POST on /a/b with application/xml should succeed on validator_XSDElementContentPlainOptMsgCodeX with valid XML1") {
     validator_XSDElementContentPlainOptMsgCodeX.validate(request("POST","/a/b","application/xml", goodXML_XSD1),response,chain)
-  }
-
-  test ("POST on /a/b with application/xml should succeed on validator_XSDElementContentPlainOptMsgCodeS with valid XML1") {
-    validator_XSDElementContentPlainOptMsgCodeS.validate(request("POST","/a/b","application/xml", goodXML_XSD1),response,chain)
   }
 
   test ("POST on /a/b with application/xml should succeed on validator_XSDElementContentPlainOpt2 with valid XML1") {
@@ -1160,10 +1110,6 @@ class ValidatorWADLPlainParamSuite extends BaseValidatorSuite {
     validator_XSDElementContentPlainOptMsgCodeX.validate(request("PUT","/a/b","application/json", goodJSON),response,chain)
   }
 
-  test ("PUT on /a/b with application/json should succeed on validator_XSDElementContentPlainOptMsgCodeS with well formed JSON") {
-    validator_XSDElementContentPlainOptMsgCodeS.validate(request("PUT","/a/b","application/json", goodJSON),response,chain)
-  }
-
   test ("PUT on /a/b with application/json should succeed on validator_XSDElementContentPlainOpt2 with well formed JSON") {
     validator_XSDElementContentPlainOpt2.validate(request("PUT","/a/b","application/json", goodJSON),response,chain)
   }
@@ -1194,10 +1140,6 @@ class ValidatorWADLPlainParamSuite extends BaseValidatorSuite {
 
   test ("POST on /c with application/json should succeed on validator_XSDElementContentPlainOptMsgCodeX with well formed JSON") {
     validator_XSDElementContentPlainOptMsgCodeX.validate(request("POST","/c","application/json", goodJSON),response,chain)
-  }
-
-  test ("POST on /c with application/json should succeed on validator_XSDElementContentPlainOptMsgCodeS with well formed JSON") {
-    validator_XSDElementContentPlainOptMsgCodeS.validate(request("POST","/c","application/json", goodJSON),response,chain)
   }
 
   test ("POST on /c with application/json should succeed on validator_XSDElementContentPlainOpt2 with well formed JSON") {
@@ -1232,10 +1174,6 @@ class ValidatorWADLPlainParamSuite extends BaseValidatorSuite {
     validator_XSDElementContentPlainOptMsgCodeX.validate(request("GET","/c"),response,chain)
   }
 
-  test ("GOT on /c should succeed on validator_XSDElementContentPlainOptMsgCodeS") {
-    validator_XSDElementContentPlainOptMsgCodeS.validate(request("GET","/c"),response,chain)
-  }
-
   test ("GOT on /c should succeed on validator_XSDElementContentPlainOpt2") {
     validator_XSDElementContentPlainOpt2.validate(request("GET","/c"),response,chain)
   }
@@ -1266,10 +1204,6 @@ class ValidatorWADLPlainParamSuite extends BaseValidatorSuite {
 
   test ("PUT on /a/b should fail with well formed XML PUT in the wrong location in validator_XSDElementContentPlainOptMsgCodeX") {
     assertResultFailed(validator_XSDElementContentPlainOptMsgCodeX.validate(request("PUT","/a/b", "application/xml", goodXML_XSD1),response,chain), 400, "Bad Content: Expecting the root element to be: tst:a")
-  }
-
-  test ("PUT on /a/b should fail with well formed XML PUT in the wrong location in validator_XSDElementContentPlainOptMsgCodeS") {
-    assertResultFailed(validator_XSDElementContentPlainOptMsgCodeS.validate(request("PUT","/a/b", "application/xml", goodXML_XSD1),response,chain), 400, "Bad Content: Expecting the root element to be: tst:a")
   }
 
   test ("PUT on /a/b should fail with well formed XML PUT in the wrong location in validator_XSDElementContentPlainOpt2") {
@@ -1304,10 +1238,6 @@ class ValidatorWADLPlainParamSuite extends BaseValidatorSuite {
     assertResultFailed(validator_XSDElementContentPlainOptMsgCodeX.validate(request("PUT","/a/b", "application/xml", badXML_Plain2),response,chain), 500, "No stepType attribute on a")
   }
 
-  test ("PUT on /a/b should fail with well formed XML PUT with missing required plain params on validator_XSDElementContentPlainOptMsgCodeS") {
-    assertResultFailed(validator_XSDElementContentPlainOptMsgCodeS.validate(request("PUT","/a/b", "application/xml", badXML_Plain2),response,chain), 500, "No stepType attribute on a")
-  }
-
   test ("PUT on /a/b should fail with well formed XML PUT with missing required plain params on validator_XSDElementContentPlainOpt2") {
     assertResultFailed(validator_XSDElementContentPlainOpt2.validate(request("PUT","/a/b", "application/xml", badXML_Plain2),response,chain), 400, "Bad Content: Expecting tst:a/@stepType")
   }
@@ -1340,9 +1270,6 @@ class ValidatorWADLPlainParamSuite extends BaseValidatorSuite {
     assertResultFailed(validator_XSDElementContentPlainOptMsgCodeX.validate(request("POST","/a/b", "application/xml", goodXML_XSD2),response,chain), 400, "Bad Content: Expecting the root element to be: tst:e")
   }
 
-  test ("POST on /a/b should fail with well formed XML POST in the wrong location in validator_XSDElementContentPlainOptMsgCodeS") {
-    assertResultFailed(validator_XSDElementContentPlainOptMsgCodeS.validate(request("POST","/a/b", "application/xml", goodXML_XSD2),response,chain), 400, "Bad Content: Expecting the root element to be: tst:e")
-  }
 
   test ("POST on /a/b should fail with well formed XML POST in the wrong location in validator_XSDElementContentPlainOpt2") {
     assertResultFailed(validator_XSDElementContentPlainOpt2.validate(request("POST","/a/b", "application/xml", goodXML_XSD2),response,chain), 400, "Bad Content: Expecting the root element to be: tst:e")
@@ -1376,10 +1303,6 @@ class ValidatorWADLPlainParamSuite extends BaseValidatorSuite {
     assertResultFailed(validator_XSDElementContentPlainOptMsgCodeX.validate(request("POST","/a/b", "application/xml", badXML_Plain1),response,chain), 501, "no stepType on e")
   }
 
-  test ("POST on /a/b should fail with well formed XML POST with missing required plain params on validator_XSDElementContentPlainOptMsgCodeS") {
-    assertResultFailed(validator_XSDElementContentPlainOptMsgCodeS.validate(request("POST","/a/b", "application/xml", badXML_Plain1),response,chain), 501, "no stepType on e")
-  }
-
   test ("POST on /a/b should fail with well formed XML POST with missing required plain params on validator_XSDElementContentPlainOpt2") {
     assertResultFailed(validator_XSDElementContentPlainOpt2.validate(request("POST","/a/b", "application/xml", badXML_Plain1),response,chain), 400, "Bad Content: Expecting tst:e/tst:stepType")
   }
@@ -1410,10 +1333,6 @@ class ValidatorWADLPlainParamSuite extends BaseValidatorSuite {
 
   test ("PUT on /a/b should fail with well formed XML that does not match schema on validator_XSDElementContentPlainOptMsgCodeX") {
     assertResultFailed(validator_XSDElementContentPlainOptMsgCodeX.validate(request("PUT","/a/b", "application/xml", goodXML),response,chain), 400)
-  }
-
-  test ("PUT on /a/b should fail with well formed XML that does not match schema on validator_XSDElementContentPlainOptMsgCodeS") {
-    assertResultFailed(validator_XSDElementContentPlainOptMsgCodeS.validate(request("PUT","/a/b", "application/xml", goodXML),response,chain), 400)
   }
 
   test ("PUT on /a/b should fail with well formed XML that does not match schema on validator_XSDElementContentPlainOpt2") {
@@ -1470,16 +1389,6 @@ class ValidatorWADLPlainParamSuite extends BaseValidatorSuite {
 
   test ("PUT on /a/b should fail with well formed XML, correct element, but does not validate against the schema in validator_XSDElementContentPlainOptMsgCodeX") {
     assertResultFailed(validator_XSDElementContentPlainOptMsgCodeX.validate(request("PUT","/a/b", "application/xml",
-                                                             <a xmlns="http://www.rackspace.com/repose/wadl/checker/step/test" stepType="foo">
-                                                                <id>21f1fcf6-bf38-11e1-878e-133ab65fcec3</id>
-                                                                <stepType>URL_FAIL</stepType>
-                                                                <even>22</even>
-                                                              </a>
-                                                            ),response,chain), 400)
-  }
-
-  test ("PUT on /a/b should fail with well formed XML, correct element, but does not validate against the schema in validator_XSDElementContentPlainOptMsgCodeS") {
-    assertResultFailed(validator_XSDElementContentPlainOptMsgCodeS.validate(request("PUT","/a/b", "application/xml",
                                                              <a xmlns="http://www.rackspace.com/repose/wadl/checker/step/test" stepType="foo">
                                                                 <id>21f1fcf6-bf38-11e1-878e-133ab65fcec3</id>
                                                                 <stepType>URL_FAIL</stepType>
