+++ pom.xml
@@ -50,11 +50,11 @@
         <scala.minor>11</scala.minor>
         <scala.patch>7</scala.patch>
         <scala.MajDotMin>${scala.major}.${scala.minor}</scala.MajDotMin>
-        <saxon-ee.version>9.7.0-15</saxon-ee.version>
+        <saxon-ee.version>9.8.0-4</saxon-ee.version>
         <scala.test.version>2.2.6</scala.test.version>
         <scala.uri.version>0.4.2</scala.uri.version>
         <junit.version>4.10</junit.version>
-        <wadl-tools.version>1.0.36</wadl-tools.version>
+        <wadl-tools.version>1.0.37</wadl-tools.version>
         <xmlsec.version>1.4.6</xmlsec.version>
         <xerces.version>2.12.1-rax</xerces.version>
         <servlet.version>3.1</servlet.version>
