+++ util/src/main/scala/com/rackspace/com/papi/compenents/checker/util/JSONConverter.scala
@@ -34,7 +34,6 @@ object JSONConverter {
   private val processor   = new Processor(false)
   private val compiler    = {
     val c = processor.newXQueryCompiler()
-    c.setLanguageVersion("3.1")
     c
   }
   private val query = "/xq/load-json.xq"
